backend/requirements.txt
fastapi[standard]
uvicorn[standard]
pymongo
python-dotenv
python-jose[cryptography] 
httpx 
python-multipart
pyJWT
pydantic-settings
prometheus-fastapi-instrumentator
prometheus-client
pytest
pytest-asyncio
mongomock
sqlalchemy
alembic
pymssql